+++ client/src/httpClient.ts
@@ -14,6 +14,7 @@ let isProcessing = false;
 let isCodeBlock = false;
 let isCodeBlockMaybeEnding = false;
 let waitingForNewline = false;
+let isCodeBlockEnded = false;
 
 /**
  * This function makes an HTTP POST request to the OpenAI API
@@ -63,8 +64,7 @@ export async function makeStreamingRequest(
   cleanup();
 
   return new Promise((resolve, reject) => {
-    let responseLogRaw: string[] = [];
-    let responseLog: string[] = [];
+    // let responseLog: string[] = []; // TODO: Only debug
     let streamedTokens = 0;
 
     let messageBuffer = new MessageBuffer();
@@ -103,12 +103,11 @@ export async function makeStreamingRequest(
             const messages = messageBuffer.process(chunkString);
 
             for (const message of messages) {
-              console.log(message);
-
               if (message === "[DONE]") {
                 stopLoading();
                 cleanup();
-                isProcessing = false;
+                isProcessing = false; // Reset state
+                isCodeBlockEnded = false; // Reset state
 
                 if (streamedTokens === 0) {
                   vscode.window.showErrorMessage(
@@ -139,9 +138,8 @@ export async function makeStreamingRequest(
                 continue;
               }
 
-              responseLog.push(token);
-
-              writeLogs(responseLog);
+              // responseLog.push(token); // TODO: only in debug
+              // writeLogs(responseLog); // TODO: only in debug
 
               if (isCodeBlock) {
                 if (checkIfCodeBlockMaybeEnding(token)) {
@@ -155,6 +153,7 @@ export async function makeStreamingRequest(
                   // Here we have gotten the confirmation that the code block
                   // is completed
                   cleanup();
+                  isCodeBlockEnded = true;
                   continue;
                 }
 
@@ -168,7 +167,6 @@ export async function makeStreamingRequest(
                 }
 
                 if (checkIfCanStream()) {
-                  console.log(`[INFO] Streaming token: ${token}`);
                   await streamCode(token, activeTextDocument);
                   streamedTokens += 1;
                 }
@@ -187,7 +185,8 @@ export async function makeStreamingRequest(
 
         res.on("end", () => {
           cleanup();
-          isProcessing = false;
+          isProcessing = false; // Reset state
+          isCodeBlockEnded = false; // Reset state
           stopLoading();
           console.log("No more data in response.");
           resolve();
@@ -286,42 +285,42 @@ function checkIfCanStream(): boolean {
  * @return {boolean} - True if a code block is starting, false otherwise.
  */
 function checkIfCodeBlockIsStarting(token: any): boolean {
-  return !isCodeBlock && token === "```";
+  return !isCodeBlock && token === "```" && !isCodeBlockEnded;
 }
 
 // TODO: Produce this only in debug mode
-function writeRawLogs(responseLogRaw: string[]) {
-  try {
-    const root = getRoot();
-    const folderPath = path.join(root, "responseLogRaw.json");
-    fs.writeFile(folderPath, JSON.stringify(responseLogRaw, null, 2), (err) => {
-      if (err) {
-        console.error("Error writing to file", err);
-      } else {
-      }
-    });
-  } catch (error) {
-    console.error(error);
-    throw error;
-  }
-}
+// function writeRawLogs(responseLogRaw: string[]) {
+//   try {
+//     const root = getRoot();
+//     const folderPath = path.join(root, "responseLogRaw.json");
+//     fs.writeFile(folderPath, JSON.stringify(responseLogRaw, null, 2), (err) => {
+//       if (err) {
+//         console.error("Error writing to file", err);
+//       } else {
+//       }
+//     });
+//   } catch (error) {
+//     console.error(error);
+//     throw error;
+//   }
+// }
 
 // TODO: Produce this only in debug mode
-function writeLogs(responseLog: string[]) {
-  try {
-    const root = getRoot();
-    const folderPath = path.join(root, "responseLog.json");
-    fs.writeFile(folderPath, JSON.stringify(responseLog, null, 2), (err) => {
-      if (err) {
-        console.error("Error writing to file", err);
-      } else {
-      }
-    });
-  } catch (error) {
-    console.error(error);
-    throw error;
-  }
-}
+// function writeLogs(responseLog: string[]) {
+//   try {
+//     const root = getRoot();
+//     const folderPath = path.join(root, "responseLog.json");
+//     fs.writeFile(folderPath, JSON.stringify(responseLog, null, 2), (err) => {
+//       if (err) {
+//         console.error("Error writing to file", err);
+//       } else {
+//       }
+//     });
+//   } catch (error) {
+//     console.error(error);
+//     throw error;
+//   }
+// }
 
 class MessageBuffer {
   private buffer: string;