+++ crates/neatcoder/src/endpoints/scaffold_project.rs
@@ -2,9 +2,15 @@ use anyhow::{anyhow, Result};
 use js_sys::{Function, JsString};
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
+use std::{
+    collections::VecDeque,
+    ops::{Deref, DerefMut},
+    path::{Path, PathBuf},
+};
 use wasm_bindgen::prelude::wasm_bindgen;
 
 use crate::{
+    consts::{CONFIG_EXTENSIONS, CONFIG_FILES},
     models::language::Language,
     openai::{
         client::OpenAI,
@@ -40,7 +46,7 @@ pub async fn scaffold_project(
     ai_params: &OpenAIParams,
     client_params: &ScaffoldParams,
     request_callback: &Function,
-) -> Result<Value> {
+) -> Result<(Value, Files)> {
     let mut prompts = Vec::new();
 
     prompts.push(OpenAIMsg {
@@ -71,7 +77,23 @@ Answer in JSON format (Do not forget to start with ```json). For each file provi
 
     process_response(&mut scaffold_json)?;
 
-    Ok(scaffold_json)
+    let mut files = Files::empty();
+
+    let src_json = scaffold_json.get("src").ok_or_else(|| {
+        anyhow!("Unable to find `src` folder in scaffold response")
+    })?;
+
+    if src_json.is_object() {
+        for (key, value) in src_json.as_object().unwrap() {
+            files.add_files(value, key, &mut PathBuf::new())?;
+        }
+    } else {
+        return Err(anyhow!("Unable to parse scaffold json."));
+    }
+
+    files.cleanup(language)?;
+
+    Ok((scaffold_json, files))
 }
 
 fn process_response(llm_response: &mut Value) -> Result<()> {
@@ -99,4 +121,133 @@ fn process_response(llm_response: &mut Value) -> Result<()> {
     }
 
     Ok(())
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Files(pub VecDeque<File>);
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct File {
+    pub name: String,
+    pub description: String,
+}
+
+impl AsRef<VecDeque<File>> for Files {
+    fn as_ref(&self) -> &VecDeque<File> {
+        &self.0
+    }
+}
+
+impl Deref for Files {
+    type Target = VecDeque<File>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl DerefMut for Files {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl Files {
+    pub fn empty() -> Files {
+        Files(VecDeque::new())
+    }
+
+    pub fn new(files: VecDeque<File>) -> Files {
+        Files(files)
+    }
+
+    pub fn add_files(
+        &mut self,
+        json: &Value,
+        current_key: &str,
+        path: &mut PathBuf,
+    ) -> Result<()> {
+        match json {
+            // If the value is an object, i.e. a map, then it means that it
+            // represents a folder. We therefore recursively call add_files itself
+            // to get to the leaf nodes, which represent the files
+            Value::Object(map) => {
+                for (key, value) in map.iter() {
+                    let mut new_path = path.join(current_key);
+                    self.add_files(value, key, &mut new_path)?;
+                }
+            }
+            // For other types of Value (Number, String, Bool, Null), consider them as leaf values.
+            _ => {
+                let full_path = path.join(current_key);
+                let filename = full_path
+                    .as_path()
+                    .to_str()
+                    .map(|s| s.to_string())
+                    .ok_or_else(|| {
+                        anyhow!("Unable to convert filename path to string :(")
+                    })?;
+
+                self.push_back(File {
+                    name: filename,
+                    description: json.to_string(),
+                });
+            }
+        }
+
+        Ok(())
+    }
+
+    pub fn cleanup(&mut self, language: &Language) -> Result<()> {
+        // Remove 'src/' prefix if it exists
+        for file in &mut self.0 {
+            if file.name.starts_with('/') {
+                file.name = file.name.trim_start_matches('/').to_string();
+            }
+
+            if file.name.starts_with("src/") {
+                file.name = file.name.trim_start_matches("src/").to_string();
+            }
+        }
+
+        // Filter out files that are configuration files, both by extension name
+        // or by filename if no extension exists
+        self.retain(|file| {
+            let path = Path::new(&file.name);
+            // This indicates that its not a file but a folder
+            if file.name.ends_with('/') {
+                return false;
+            }
+
+            if let Some(file_name) = path.file_name().and_then(|n| n.to_str()) {
+                if CONFIG_FILES.contains(&file_name) {
+                    return false;
+                }
+            }
+            if let Some(extension) =
+                path.extension().and_then(|ext| ext.to_str())
+            {
+                if CONFIG_EXTENSIONS.contains(&extension) {
+                    return false;
+                }
+            }
+            true
+        });
+
+        if !language.is_custom() {
+            let default_extension = language.language.default_extension();
+
+            // If GPT forgot to add the file extensions we add them here..
+            if let Some(default_ext) = default_extension {
+                for file in &mut self.0 {
+                    let path = Path::new(&file.name);
+                    if path.extension().is_none() {
+                        file.name.push_str(default_ext);
+                    }
+                }
+            }
+        }
+
+        Ok(())
+    }
 }